+++ docs/CHANGELOG.md
@@ -70,6 +70,8 @@ Version 1.9.0 (not released yet)
 - Add `spectate_mode_follow_killer` command (when player you are spectating dies, spectate their killer)
 - Add current server player count to scoreboard
 - Enable commands `drop_clutter`, `drop_entity`, `drop_item`, `teleport`, `pcollide` in single player
+- Add missing help strings for builtin cheat commands
+- Make `spectate` command set camera to free look when issued in single player
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------